Fix linking against older versions of Qt
tst_perfparser.cpp:(.text._ZN14TestPerfParser24testCppInliningNoOptionsEv[_ZN14TestPerfParser24testCppInliningNoOptionsEv]+0x2b1):
undefined reference to `bool QTest::qCompare<unsigned long long, int>(unsigned long long const&, int const&, char const*, char const*, char const*, int)'

+++ tests/integrationtests/tst_perfparser.cpp
@@ -145,8 +145,8 @@ private slots:
         QVERIFY(!m_topDownData.root.children.isEmpty());
 
         // we don't know the on/off CPU time
-        QCOMPARE(m_summaryData.onCpuTime, 0);
-        QCOMPARE(m_summaryData.offCpuTime, 0);
+        QCOMPARE(m_summaryData.onCpuTime, quint64(0));
+        QCOMPARE(m_summaryData.offCpuTime, quint64(0));
     }
 
     void testCppInliningCallGraphDwarf()
